Release Instance Files
Instance files are a new method of tracking all the data we have which pertains to the creatures/objects within all dungeons and raids.
Each file corresponds to an individual instance and the first three digits of the filename is the official number of that map.

The files currently contain work mostly from X-Savior, Malcrom, Tobschinski, Cala, Saltgurka, AnonXS, Killerwife, and MantisLord (and many others before us in CMaNGOS/MaNGOS).

One of the main goals of creating these files originally was organization.
With the format we've used, it becomes an incredibly simplistic process to re-guid all dungeons/raids at once with no conflicts (just change one number per file).
We've also found it's easier for developers who are scripting a particular instance if they can locate and work with all relevant data in one centralized location.

Another big objective of switching to this design was to minimize our diff with Classic-DB as well as WotLK-DB wherever possible.
When generating the files for vanilla dungeons/raids, Classic-DB was used as the source and any TBC specific data was added in and marked with a note/comment of "TBC+ ONLY".

With this approach, it becomes very clear which data is shared between expansions and which data changed over time.
The theory is, the Classic-DB team should be able to use the exact same files we have here, with the TBC+ ONLY sections removed.
The files should also apply 1:1 for WotLK-DB with a few exceptions such as Naxxramas.